createListing: update log & move whitelist function :
