+++ contracts.json
@@ -1,1 +1,1 @@
-{"goerli":{"nftContract":"0x22411b480425fe6e627fdf4d1b6ac7f8567314ada5617a0a6d8ef3e74b69436","messaging":"0x2c3d3e0c37d29364a13ba8cff046e7bc5624655a72526961876a1c8bb3f63c8","executor":"0x73148536f8ea9546e92761d11515548cc433df46883d5ee98871a6f63a0bbbc","orderbook":"0x66f1e6acf9bdbd23837b2eea271430298b355c506978edb132737e7fcb6b310"},"sepolia":{},"mainnet":{},"dev":{"messaging":"0x1fed294721e45a912cd64ec6aba8b9e902b6f54fd516547ca740805607d47cd","executor":"0x4a69823e775f40d8837c62cf9e239c3da4ab65888dd6d8f19e6f9b11771e460","nftContract":"0x664b395332553a06c12f6164f11544b92b1ec308fa6c360f01f9cde1cc32ac2","eth":"0x4007c4afac7b79c2119c1d417e3106294cb28d92bdfba26474e8e250b1e91c9","orderbook":"0x1b59160f5593a11ea887b063d541a42f758d82e0b7295763e3104e09057ab5"}}
+{"goerli":{"nftContract":"0x22411b480425fe6e627fdf4d1b6ac7f8567314ada5617a0a6d8ef3e74b69436","messaging":"0x2c3d3e0c37d29364a13ba8cff046e7bc5624655a72526961876a1c8bb3f63c8","executor":"0x73148536f8ea9546e92761d11515548cc433df46883d5ee98871a6f63a0bbbc","orderbook":"0x66f1e6acf9bdbd23837b2eea271430298b355c506978edb132737e7fcb6b310"},"sepolia":{},"mainnet":{},"dev":{"messaging":"0x53a0b0ca4b9d248dff574299c98340b67197f7ffc79df9afef66d01e83022c6","executor":"0x32cee2469074dabb6ee8dcce24df08d79e9ffd5f978163fbdd0704bce269bab","nftContract":"0x305bf776b8976a9709dde4cf307aebed4060e201ab3fca7154ca6db386f13c","eth":"0x144a0c8ae6e665dc3652e164ba2881eda87ad9cb5dcc93d30ca6d61b818014c","orderbook":"0x62a58d7c4680d2fd5b96ac36a5f771a4127cdcedc89fc0a5932c32ea2803f59"}}
