+++ crates/solis/messaging.local.json
@@ -1,9 +1,9 @@
 {
   "chain": "starknet",
   "rpc_url": "http://127.0.0.1:5050",
-  "contract_address": "0x1fed294721e45a912cd64ec6aba8b9e902b6f54fd516547ca740805607d47cd",
-  "sender_address": "0x2d71e9c974539bb3ffb4b115e66a23d0f62a641ea66c4016e903454c8753bbc",
-  "private_key": "0x33003003001800009900180300d206308b0070db00121318d17b5e6262150b",
+  "contract_address": "0x53a0b0ca4b9d248dff574299c98340b67197f7ffc79df9afef66d01e83022c6",
+  "sender_address": "0x6162896d1d7ab204c7ccac6dd5f8e9e7c25ecd5ae4fcb4ad32e57786bb46e03",
+  "private_key": "0x1800000000300000180000000000030000000000003006001800006600",
   "interval": 2,
   "from_block": 0
 }
